add list of guessed letters, fix issue with displaying multiple times on loading.

+++ hangman.rb
@@ -1,7 +1,7 @@
 require 'yaml'
 
 class Game
-  attr_accessor :word, :display_word, :guesses, :status, :player, :playing, :command
+  attr_accessor :word, :display_word, :guesses, :status, :player, :playing, :command, :guessed
   def initialize
     @wordlist = File.read('wordlist.txt')
     @word = select_word
@@ -9,6 +9,7 @@ class Game
     @guesses = 7
     @player = Player.new(self)
     @playing = true
+    @guessed = []
   end
 
   def select_word
@@ -29,11 +30,12 @@ class Game
   end
 
   def display
-    puts @display_word.split('').join(' ')
-    puts "#{@guesses} guesses remaining." if @playing == true
+    puts "#{@display_word.split('').join(' ')}\tGuessed letters: #{@guessed.join}"
+    puts "#{@guesses} guesses remaining.\n\n" if @playing == true
   end
 
   def save_game
+    puts "Saving..."
     yaml = YAML::dump(self)
     save_file = File.new('save_file.yaml', 'w')
     save_file.write(yaml)
@@ -42,7 +44,6 @@ class Game
 end
 
 class Player
-
   def initialize(game)
     @game = game
     @word = game.word
@@ -66,7 +67,7 @@ class Player
       @game.save_game
       @game
     elsif @choice == 'load'
-      @game = load_game
+      load_game || @game
     elsif @choice == 'quit'
       exit
     else
@@ -79,14 +80,14 @@ class Player
   end
 
   def check_guess
+    @game.guessed << @choice
     if @word =~ /#{@choice}/
-      puts "Correct!\n"
+      puts "Correct!\n\n"
       positions = []
       @word.split('').each_with_index { |letter, index| positions << index if letter == @choice}        
-      positions.each {|index| @game.display_word[index] = "#{@choice}"}
-    
+      positions.each {|index| @game.display_word[index] = "#{@choice}"}   
     else
-      puts "Wrong.\n"
+      puts "Wrong.\n\n"
       @game.guesses -= 1
     end
   end
@@ -106,15 +107,20 @@ def play_game
   end_game
 end
 
+def new_game
+  game = Game.new
+  game.display
+  return game
+end
+
 def new_or_load
   begin
     choice = gets.chomp.strip.downcase
     if choice == 'new'
-      game = Game.new
-      game.display
-      return game
+      new_game
     elsif choice == 'load'
-      load_game
+      game = load_game
+      game ? game : new_game
     else
       raise ArgumentError, "Invalid choice. Please type 'new' or 'load'"
     end
@@ -125,11 +131,16 @@ def new_or_load
 end
 
 def load_game
-  save_file = File.open('save_file.yaml', 'r')
-  yaml = save_file.read
-  game = YAML::load(yaml)
-  game.display
-  return game
+  if File.exist?('save_file.yaml')
+    save_file = File.open('save_file.yaml', 'r')
+    yaml = save_file.read
+    game = YAML::load(yaml)
+    game.display
+    return game
+  else
+    puts "No save file exists."
+    return false
+  end
 end
 
 def end_game
